move JsonResponseContract to parent class
